ci: add zizmor to interface-tests-v0 branch (#241)
This PR adds a job to run `zizmor` in our CI. The job is exactly the
same as on `main`. We also copy the `zizmor.yml` config file from
`main`.

While we're at it, I added the conventional PR titles workflow to this
branch as well, just copied directly from `main`.